title: Card 컴포넌트 분리
- Card 컴포넌트는 공통 적용부분만 분리
- TeamMemberCard의 모달 state 직접관리
- 기타 코드 정리
