+++ osanReactStudy/src/App.jsx
@@ -1,5 +1,4 @@
-import { useState } from "react";
-import TeamMemberCard from "./components/TeamMemberCard";
+import { TeamMemberCard } from "./components/Card";
 
 const memberList = [
   {
@@ -31,11 +30,14 @@ function App() {
         <h1 className="text-2xl font-bold">Team Members</h1>
       </section>
       <section id="team_members" className="container-fluid mx-auto p-4">
-        <div className="grid sm:grid-cols-2 md:grid-cols-3 xl:grid-cols-4
-          gap-3 md:gap-4">
+        <div
+          className="grid sm:grid-cols-2 md:grid-cols-3 xl:grid-cols-4
+          gap-3 md:gap-4"
+        >
           {memberList.map((member, index) => (
             <TeamMemberCard key={index} member={member} />
           ))}
+          <TeamMemberCard />
         </div>
       </section>
     </main>
